requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.5.0
transformers==4.15.0
pillow==10.4.0
gunicorn==21.2.0 
